packages/ingestion/pyproject.toml
[project]
name = "evergreen-ingestion"
version = "0.1.0"
description = "Evergreen M365 Roadmap Ingestion Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.30",
    "google-generativeai>=0.8",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
